Move `test_targets_by_name` to scarb-api
